DevFest25-WS/P3-AgentTeams/pyproject.toml
[project]
name = "p3-agentteams"
version = "0.1.0"
description = "Part 3: Building Agent Teams - The Specialist Pattern"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "google-generativeai",
    "python-dotenv",
]

